First try fixing red containers and stacks > 5

+++ libraries/_Logic/_Logic.h
@@ -9,7 +9,7 @@ class _Logic{
         int *stacks;
         uint8_t blue_boxes, green_boxes;
         uint8_t lastStack, currentLevel, lastLevel;
-        uint8_t firstRed;      
+        int firstRed;      
         char lastColor, c_serial;
     
     public:
